+++ test/support/recovery_mocks.exs
@@ -1,0 +1,4 @@
+require Mox
+
+Mox.defmock(Tinkex.Recovery.RestMock, for: Tinkex.Recovery.RestBehaviour)
+Mox.defmock(Tinkex.Recovery.ServiceClientMock, for: Tinkex.Recovery.ServiceClientBehaviour)