+++ test/test_helper.exs
@@ -18,6 +18,8 @@ System.put_env("TINKER_TELEMETRY", "0")
 
 Code.require_file("support/http_case.ex", __DIR__)
 Code.require_file("support/api_worker.ex", __DIR__)
+Code.require_file("support/recovery_stubs.exs", __DIR__)
+Code.require_file("support/recovery_mocks.exs", __DIR__)
 
 telemetry_handler = "supertester-phase2"
 event_name = [:supertester, :concurrent, :scenario, :stop]